+++ lib/queue.js
@@ -7,9 +7,9 @@ var querystring = require('querystring');
 var crypto      = require('crypto');
 var events      = require('events');
 var util        = require('util');
-var libxml      = require('libxmljs');
 var agent       = require('./agent');
 var utils       = require('./utils');
+var xml         = require('./xml');
 
 /*
  * Azure storage service version
@@ -484,53 +484,11 @@ Queue.prototype.request = function request(method, path, query, headers, data) {
           return res;
         }
 
-        // Parse payload for error message and code
-        var message, code, detail;
-        try {
-          var xml = libxml.parseXml(res.payload);
-
-          // Find error message
-          message = xml.get('/Error/Message');
-          if (message) {
-            message = message.text();
-          }
-
-          // Find error code
-          code = xml.get('/Error/Code');
-          if (code) {
-            code = code.text();
-          }
-
-          detail = xml.get('/Error/AuthenticationErrorDetail');
-          if (detail) {
-            detail = detail.text();
-          }
-        } catch (e) {
-          // Ignore parsing errors
-        }
-
-        // Find default message and code
-        if (!message) {
-          message = "No error message given, in payload '" + res.payload + "'";
-        }
-        if (!code) {
-          if (500 <= res.statusCode && res.statusCode < 600) {
-            code = 'InternalErrorWithoutCode';
-          } else {
-            code = 'ErrorWithoutCode';
-          }
-        }
-
-        // Construct error object
-        var err         = new Error(message);
-        err.name        = code + 'Error';
-        err.code        = code;
-        err.statusCode  = res.statusCode;
-        err.detail      = detail;
-        err.payload     = res.payload;
-
+        // Parse error code
+        var err = xml.queueParseError(res);
+        err.retries = retry;
         debug("Error code: %s for %s %s on retry: %s",
-              code, method, path, retry);
+              err.code, method, path, retry);
 
         // Throw the constructed error
         throw err;
@@ -581,50 +539,7 @@ Queue.prototype.listQueues = function listQueues(options) {
   if (options.metadata)   query.include     = 'metadata';
 
   // Send request with retry policy
-  return this.request('GET', '/', query, {}).then(function(res) {
-    // Get results
-    var xml     = libxml.parseXml(res.payload);
-    var queues  = xml.get('/EnumerationResults/Queues').childNodes();
-
-    // Construct results
-    var result = {
-      queues: queues.map(function(queue) {
-        var metadata = undefined;
-        var metaNode = queue.get('Metadata');
-        if (metaNode) {
-          metadata = {};
-          metaNode.childNodes().forEach(function(node) {
-            metadata[node.name()] = node.text();
-          });
-        }
-        return {
-          name:     queue.get('Name').text(),
-          metadata: metadata
-        };
-      })
-    };
-
-    // Get Marker, Prefix, MaxResults and NextMarker, if present
-    var marker = xml.get('/EnumerationResults/Marker');
-    if (marker) {
-      result.marker = marker.text();
-    }
-    var prefix = xml.get('/EnumerationResults/Prefix');
-    if (prefix) {
-      result.prefix = prefix.text();
-    }
-    var maxResults = xml.get('/EnumerationResults/MaxResults');
-    if (maxResults) {
-      result.maxResults = parseInt(maxResults.text());
-    }
-    var nextMarker = xml.get('/EnumerationResults/NextMarker');
-    if (nextMarker ) {
-      result.nextMarker = nextMarker.text();
-    }
-
-    // Return result
-    return result;
-  });
+  return this.request('GET', '/', query, {}).then(xml.queueParseListQueues);
 };
 
 // TODO: Implement someday when we need it:
@@ -844,20 +759,7 @@ Queue.prototype.peekMessages = function peekMessages(queue, options) {
   }
 
   // Send request with retry policy
-  return this.request('GET', path, query, {}).then(function(res) {
-    // Get results
-    var xml   = libxml.parseXml(res.payload);
-    var msgs  = xml.get('/QueueMessagesList').childNodes();
-    return msgs.map(function(msg) {
-      return {
-        messageId:        msg.get('MessageId').text(),
-        insertionTime:    new Date(msg.get('InsertionTime').text()),
-        expirationTime:   new Date(msg.get('ExpirationTime').text()),
-        dequeueCount:     parseInt(msg.get('DequeueCount').text()),
-        messageText:      msg.get('MessageText').text()
-      };
-    });
-  });
+  return this.request('GET', path, query, {}).then(xml.queueParsePeekMessages);
 };
 
 /**
@@ -908,22 +810,7 @@ Queue.prototype.getMessages = function getMessages(queue, options) {
   }
 
   // Send request with retry policy
-  return this.request('GET', path, query, {}).then(function(res) {
-    // Get results
-    var xml   = libxml.parseXml(res.payload);
-    var msgs  = xml.get('/QueueMessagesList').childNodes();
-    return msgs.map(function(msg) {
-      return {
-        messageId:        msg.get('MessageId').text(),
-        insertionTime:    new Date(msg.get('InsertionTime').text()),
-        expirationTime:   new Date(msg.get('ExpirationTime').text()),
-        dequeueCount:     parseInt(msg.get('DequeueCount').text()),
-        messageText:      msg.get('MessageText').text(),
-        popReceipt:       msg.get('PopReceipt').text(),
-        timeNextVisible:  new Date(msg.get('TimeNextVisible').text())
-      };
-    });
-  });
+  return this.request('GET', path, query, {}).then(xml.queueParseGetMessages);
 };
 
 /**
